+++ crates/tauri/src/lib.rs
@@ -18,7 +18,6 @@
 //! - **tracing**: Enables [`tracing`](https://docs.rs/tracing/latest/tracing) for window startup, plugins, `Window::eval`, events, IPC, updater and custom protocol request handlers.
 //! - **test**: Enables the [`mod@test`] module exposing unit test helpers.
 //! - **objc-exception**: Wrap each msg_send! in a @try/@catch and panics if an exception is caught, preventing Objective-C from unwinding into Rust.
-//! - **linux-ipc-protocol**: Use custom protocol for faster IPC on Linux. Requires webkit2gtk v2.40 or above.
 //! - **linux-libxdo**: Enables linking to libxdo which enables Cut, Copy, Paste and SelectAll menu items to work on Linux.
 //! - **isolation**: Enables the isolation pattern. Enabled by default if the `app > security > pattern > use` config option is set to `isolation` on the `tauri.conf.json` file.
 //! - **custom-protocol**: Feature managed by the Tauri CLI. When enabled, Tauri assumes a production environment instead of a development one.
@@ -389,7 +388,6 @@ pub struct Context<R: Runtime> {
   #[cfg(all(desktop, feature = "tray-icon"))]
   pub(crate) tray_icon: Option<image::Image<'static>>,
   pub(crate) package_info: PackageInfo,
-  pub(crate) _info_plist: (),
   pub(crate) pattern: Pattern,
   pub(crate) runtime_authority: RuntimeAuthority,
   pub(crate) plugin_global_api_scripts: Option<&'static [&'static str]>,
@@ -503,7 +501,6 @@ impl<R: Runtime> Context<R> {
     default_window_icon: Option<image::Image<'static>>,
     app_icon: Option<Vec<u8>>,
     package_info: PackageInfo,
-    info_plist: (),
     pattern: Pattern,
     runtime_authority: RuntimeAuthority,
     plugin_global_api_scripts: Option<&'static [&'static str]>,
@@ -518,7 +515,6 @@ impl<R: Runtime> Context<R> {
       #[cfg(all(desktop, feature = "tray-icon"))]
       tray_icon: None,
       package_info,
-      _info_plist: info_plist,
       pattern,
       runtime_authority,
       plugin_global_api_scripts,